pyproject.toml
[tool.poetry]
name = "krylovilya"
version = "0.1.0"
description = ""
authors = ["krylovilya1 <krylovilya1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2.8"
django-ckeditor = "^6.1.0"
django-constance = "^2.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
